+++ cmd/unload.go
@@ -4,6 +4,7 @@ import (
 	"envManager/environment"
 	"envManager/helper"
 	"envManager/secretsStorage"
+	"fmt"
 	"github.com/spf13/cobra"
 	"strings"
 )
@@ -14,16 +15,23 @@ var unloadCmd = &cobra.Command{
 	Short:             "Unload profiles",
 	Long:              `Unload one or more profiles from this shell's environment`,
 	Run:               runUnload,
-	Args:              cobra.MinimumNArgs(1),
 	ValidArgsFunction: CompleteProfiles,
 	PreRun:            InitConfig,
 }
 
-func runUnload(_ *cobra.Command, args []string) {
+func runUnload(cmd *cobra.Command, args []string) {
 	registry := secretsStorage.GetRegistry()
 	env := environment.NewEnvironment()
 	env.Load()
 	loadedProfiles := strings.Split(env.GetCurrent(envManagerLoadedProfilesName, ""), ",")
+	unloadAllFlagSet, _ := cmd.Flags().GetBool("all")
+	if unloadAllFlagSet {
+		// pretend all loaded profiles were listed as args
+		args = loadedProfiles
+	} else if len(args) == 0 {
+		// no --all flag and no profile name specified
+		fmt.Println("You must specify at least one profile to unload")
+	}
 	for i := 0; i < len(args); i++ {
 		profile, err := registry.GetProfile(args[i])
 		cobra.CheckErr(err)
@@ -38,4 +46,5 @@ func runUnload(_ *cobra.Command, args []string) {
 
 func init() {
 	rootCmd.AddCommand(unloadCmd)
+	unloadCmd.Flags().Bool("all", false, "Select all currently loaded profiles for unloading")
 }